Model validations
Was added the respective validations for each model, the validation was
as follow:

Was validated that each field of each model mathces with the regexp,
that, so the string must have letters, numbers and special characters,
if applicable.

Finally was changed the structure in each controller to render the
response.