API versioning
Was added the namespace :api and inside the namespace :v1, and then the
resources.

To do this accessible, wass necessary add the Api::V1:: befor the name
of each controller, after that, is necessary create the api directory
and inside the v1 directory, inside this directory, goes all the
controllers.